sync the strndup fallback parameter type
svn-id: r7131

+++ gemrb/includes/globals.h
@@ -178,7 +178,7 @@ GEM_EXPORT int EARelation(Scriptable *a, Actor *b);
 GEM_EXPORT bool dir_exists(const char* path);
 GEM_EXPORT int strlench(const char* string, char ch);
 #ifndef HAVE_STRNDUP
-GEM_EXPORT char* strndup(const char* s, unsigned int l);
+GEM_EXPORT char* strndup(const char* s, size_t l);
 #endif
 
 #ifndef WIN32
